nit: type upcast for explicitness

+++ contracts/plugins/assets/FiatCollateral.sol
@@ -188,7 +188,7 @@ contract FiatCollateral is ICollateral, Asset {
         if (status_ == CollateralStatus.SOUND) {
             _whenDefault = NEVER;
         } else if (status_ == CollateralStatus.IFFY) {
-            uint256 sum = block.timestamp + delayUntilDefault;
+            uint256 sum = block.timestamp + uint256(delayUntilDefault);
             if (sum >= NEVER) _whenDefault = NEVER;
             else if (sum < _whenDefault) _whenDefault = uint48(sum);
             // else: no change to _whenDefault
